lab1/requirements.txt
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.7
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
scipy==1.16.2
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
typing_extensions==4.15.0
